refactor(persis): relocate filesession under internal/persis/file/
Move filesession into the persis/file/ subtree alongside dagrun and proc.
The store keeps its nested layout ({userID}/{sessionID}.json), which is
file-backend specific and not layout-compatible with store/session.go
(flat {sessionID}.json). Consumer imports use the filesession alias so
the caller code is unchanged.
